requirements.txt
# Core Telegram Bot Framework
telethon>=1.28.0

# Environment and Configuration
python-dotenv>=1.0.0

# HTTP Client for API calls (OpenAI, Grok, etc.)
aiohttp>=3.8.0

# Google Cloud and Firebase
google-cloud-firestore>=2.13.0
firebase-admin>=6.2.0

# Timezone handling
pytz>=2023.3

# Data Science and ML
numpy>=1.24.0
scikit-learn>=1.3.0

# Memory/Vector Database Client
mem0>=0.1.0

# Async support (usually included with Python 3.7+, but explicit for clarity)
asyncio-helpers>=0.2.0

# JSON handling (built-in, but some projects need enhanced JSON)
# json is built-in, no need to include

# Additional utilities
requests>=2.31.0 